+++ src/app/pages/address/address.component.ts
@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-address',
@@ -21,6 +22,13 @@ export class AddressComponent implements OnInit {
   // ✅ FIX 1
   ngOnInit() {
     this.loadAddresses();
+    const editData = localStorage.getItem("editAddress");
+
+  if (editData) {
+    this.newAddress = JSON.parse(editData);
+    this.editMode = true;
+    this.editId = this.newAddress.id;
+  }
   }
 
   loadAddresses() {
@@ -63,7 +71,14 @@ saveAddress() {
     // 🔥 UPDATE API
     this.http.put(`http://localhost:5000/api/address/update/${this.editId}`, payload)
       .subscribe(() => {
-        alert("Address Updated ✏️");
+        Swal.fire({
+  icon: 'success',
+  title: 'Updated!',
+  text: 'Address updated successfully ✏️',
+  showConfirmButton: false,
+  timer: 1800
+});
+        localStorage.removeItem("editAddress");
         this.resetForm();
         this.loadAddresses();
       });
@@ -72,7 +87,13 @@ saveAddress() {
     // ➕ ADD API
     this.http.post('http://localhost:5000/api/address/add', payload)
       .subscribe(() => {
-        alert("Address Added ✅");
+       Swal.fire({
+  icon: 'success',
+  title: 'Added!',
+  text: 'New address saved 📍',
+  showConfirmButton: false,
+  timer: 1800
+});
         this.resetForm();
         this.loadAddresses();
       });
@@ -93,13 +114,38 @@ resetForm() {
   this.newAddress = {};
   this.editMode = false;
   this.editId = null;
+  localStorage.removeItem("editAddress");
 }
-  deleteAddress(id: number) {
-    this.http.delete(`http://localhost:5000/api/address/delete/${id}`)
-      .subscribe(() => {
-        this.loadAddresses();
-      });
-  }
+
+deleteAddress(id: number) {
+  Swal.fire({
+    title: 'Are you sure?',
+    text: 'This address will be deleted!',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#d33',
+    cancelButtonColor: '#3085d6',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+
+      this.http.delete(`http://localhost:5000/api/address/delete/${id}`)
+        .subscribe(() => {
+          
+          Swal.fire({
+            icon: 'success',
+            title: 'Deleted!',
+            text: 'Address removed 🗑️',
+            timer: 1500,
+            showConfirmButton: false
+          });
+
+          this.loadAddresses();
+        });
+
+    }
+  });
+}
 
   // defaulyt address
   setDefaultAddress(selected: any) {
